Bug Fixes & Improvements
Removed 'if (true)'s for true is always = to true. (Unneeded)
Fixed shop prices, You could buy stuff and sell for more.
Fixed selling vecs that where close to the shop selling point but not
close to you
Fixed selling vecs that are the same (type) but not the same "king"
(Example: RU Bomber = typeOf Su34 & RU AA = typeOf Su34)